[fix]
[fix] VSTE-485 check export to root level on windows does not work
[update] remove warnings

+++ source/linkcontroller.cpp
@@ -33,7 +33,7 @@ LinkController::LinkController ()
 	fs.seek (0, SeekableStream::SeekMode::kSeekSet);
 
 	Buffer<char> buffer (fileSize);
-	fs.readRaw (buffer.data (), buffer.size ());
+	fs.readRaw (buffer.data (), static_cast<uint32_t> (buffer.size ()));
 
 	try
 	{
